back/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==10.2.0
python-decouple==3.8
psycopg2-binary==2.9.9
drf-spectacular==0.27.1
requests==2.31.0
